+++ src/app.ts
@@ -3,12 +3,11 @@ import * as express from "express";
 import * as nunjucks from "nunjucks";
 import * as path from "path";
 import * as createError from "http-errors";
-
 import {appRouter} from "./routes/routes";
 import {ERROR_SUMMARY_TITLE} from "./model/error.messages";
 import * as pageURLs from "./model/page.urls";
-import sessionMiddleware from "./session/middleware";
-import cookieMiddleware from "./session/middleware/cookie";
+import chSessionLoader from "./session/middleware/ch_session";
+import ptfSessionLoader from "./session/middleware/ptf_session";
 import {PIWIK_SITE_ID, PIWIK_URL} from "./properties";
 import authenticate from "./authentication/middleware/index";
 import logger from "./logger";
@@ -37,8 +36,8 @@ app.enable("trust proxy");
 app.use(express.json());
 app.use(express.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(cookieMiddleware);
-app.use(sessionMiddleware);
+app.use(chSessionLoader);
+app.use(ptfSessionLoader);
 
 app.use(`${pageURLs.PROMISE_TO_FILE}/*`, authenticate);
