+++ src/session/middleware/ptf_session/index.ts
@@ -1,6 +1,6 @@
 import { NextFunction, Request, Response } from "express";
-import * as keys from "../keys";
-import * as sessionService from "../../services/session.service";
+import * as keys from "../../keys";
+import * as sessionService from "../../../services/session.service";
 
 export default async (req: Request, res: Response, next: NextFunction) => {
   const existing = req.chSession.data[keys.PTF_SESSION];
